+++ deep_groove_project/requirements.txt
@@ -1,44 +1,56 @@
+autoPyTorch==0.0.2
 certifi==2022.12.7
-charset-normalizer==3.1.0
-cmake==3.26.0
+charset-normalizer==2.1.1
+ConfigSpace==0.6.1
 contourpy==1.0.7
 cycler==0.11.0
-filelock==3.10.0
-fonttools==4.39.2
+Cython==0.29.34
+fasteners==0.18
+filelock==3.9.0
+fonttools==4.39.3
+hpbandster==0.7.4
 idna==3.4
-importlib-resources==5.12.0
+imbalanced-learn==0.10.1
+imblearn==0.0
 Jinja2==3.1.2
+joblib==1.2.0
 kiwisolver==1.4.4
-lit==15.0.7
+liac-arff==2.5.0
 MarkupSafe==2.1.2
 matplotlib==3.7.1
-mpmath==1.3.0
+minio==7.1.14
+more-itertools==9.1.0
+mpmath==1.2.1
+netifaces==0.11.0
 networkx==3.0
-numpy==1.24.2
-nvidia-cublas-cu11==11.10.3.66
-nvidia-cuda-cupti-cu11==11.7.101
-nvidia-cuda-nvrtc-cu11==11.7.99
-nvidia-cuda-runtime-cu11==11.7.99
-nvidia-cudnn-cu11==8.5.0.96
-nvidia-cufft-cu11==10.9.0.58
-nvidia-curand-cu11==10.2.10.91
-nvidia-cusolver-cu11==11.4.0.1
-nvidia-cusparse-cu11==11.7.4.91
-nvidia-nccl-cu11==2.14.3
-nvidia-nvtx-cu11==11.7.91
-packaging==23.0
-pandas==1.5.3
-Pillow==9.4.0
+numpy==1.24.1
+openml==0.13.1
+packaging==23.1
+pandas==2.0.0
+patsy==0.5.3
+Pillow==9.3.0
+protobuf==4.22.3
+psutil==5.9.4
+pyarrow==11.0.0
+pynisher==1.0.5
 pyparsing==3.0.9
+Pyro4==4.82
 python-dateutil==2.8.2
-pytz==2022.7.1
-requests==2.28.2
+pytz==2023.3
+pywin32==306
+requests==2.28.1
+scikit-learn==1.2.2
+scipy==1.10.1
+serpent==1.41
 six==1.16.0
+statsmodels==0.13.5
 sympy==1.11.1
-torch==2.0.0
-torchaudio==2.0.1
-torchvision==0.15.1
-triton==2.0.0
-typing-extensions==4.5.0
-urllib3==1.26.15
-zipp==3.15.0
+tensorboard-logger==0.1.0
+threadpoolctl==3.1.0
+torch==2.0.0+cu117
+torchaudio==2.0.1+cu117
+torchvision==0.15.1+cu117
+typing_extensions==4.4.0
+tzdata==2023.3
+urllib3==1.26.13
+xmltodict==0.13.0
